+++ test/correctness/PySys_internal_197/Input/Test/pysystest.py
@@ -1,0 +1,13 @@
+__pysys_title__   = r""" A test """ 
+import os, sys, math, shutil, glob
+
+import pysys.basetest, pysys.mappers
+from pysys.constants import *
+
+class PySysTest(pysys.basetest.BaseTest):
+
+	def execute(self):
+		pass
+		
+	def validate(self):
+		pass